+++ crates/core/src/fs/mod.rs
@@ -12,7 +12,7 @@ mod save;
 #[cfg(feature = "gui")]
 mod theme;
 
-pub use addon::{delete_addons, install_addon};
+pub use addon::{delete_addons, delete_saved_variables, install_addon};
 pub use save::PersistentData;
 #[cfg(feature = "gui")]
 pub use theme::load_user_themes;